Add packaged and native environments
Our sugar-build environment is best for developing Sugar itself, and is
not needed for developing activities.

A simpler environment for developing activities is added; using the
latest downstream packages on Debian or Ubuntu, and the Remote Desktop
Protocol (RDP) for wrapping the session in a window.

A native environment is also added, but this requires much expertise to
maintain.