Avoid activities.sugarlabs.org for Python 3 bundles
Restrict activities.sugarlabs.org to Python 2, JavaScript, and
activities compatible with Sugar 0.112.

Sugar 0.113 adds Python 3 toolkit support, which enables porting of
Python 2 activities to Python 3.

Activities written with Python 3 will not work on Sugar 0.112 and
earlier.

Automatic update in all currently released versions of Fedora SoaS, and
manual update using My Setting - Software Update in Fedora, Debian, and
Ubuntu, will break an activity if a bundle is released.

Sugar Activity Library is coded to not offer activities for
incompatible versions of Sugar, but the feature is not working.
https://github.com/sugarlabs/sugar/issues/818

Oversight board has not decided to extend security maintenance of
activities.sugarlabs.org.