Add Port to TelepathyGLib for Port to Python 3
Guide for porting to Python 3 did not have details of porting to TelepathyGLib.

Explain why constants are being renamed; for new developers it is not obvious why we are renaming the Telepathy D-Bus API constants.  We're doing it to minimise change and ease review.  Later porting can change to the TelepathyGLib names if that is helpful, but it isn't a goal of porting to Python 3.

Test and fix collaboration before and after porting; some developers have launched into porting to TelepathyGLib without testing if collaboration is actually needed and working.  Their work was
wasted.

Mention the fedora18 branch practice; some activities will continue to be maintained in Python 2 as well as Python 3, for selected systems.  Identify the branch name used.

Recommend Toolkit 0.114 as it contains a patch to build for either Python 2 or Python 3.

Add test for successful installation using Python 3.